nearly finished moving synth code
