+++ ngapp/src/lib/synth.ts
@@ -1,9 +1,8 @@
 import PlaybackHandle from "./synth/playbackHandle";
-import showButtons from "./synth/showButtons";
 import Buttons from "./synth/buttons";
 
 const synth = {
-  PlaybackHandle, showButtons, Buttons,
+  PlaybackHandle, Buttons,
 }
 
 export default synth;
